style changes and minor refactoring
