--- crack_test.txt
@@ -1,1 +1,0 @@
-(:Nv;bYy?CF)R:PnRc
