--- encrypt.txt
@@ -1,1 +1,0 @@
-x<#d4Y<g7\63;<%V;Z
