--- encrypted.txt
@@ -1,1 +1,0 @@
-R43YYD#bb.vyv45QvE
