+++ lib/key_gen.rb
@@ -1,6 +1,5 @@
 require 'Date'
 require 'Time'
-require 'pry'
 
 class KeyGen
   attr_reader   :random_key,
@@ -10,16 +9,16 @@ class KeyGen
     @random_key  = 5.times.map { Random.rand(10) }
     @date_string = Date.today.strftime("%d%m%y")
   end
+  
+  def key_normalizer(key)
+    key = key.split("").map { |number| number.to_i }
+    key_offset(key)
+  end
 
   def key_offset(key = random_key)
     key.map.with_index do |element, index|
       "#{element}#{key[index + 1]}".to_i
     end.shift(4)
-  end
-
-  def key_normalizer(key)
-    key = key.split("").map { |number| number.to_i }
-    key_offset(key)
   end
 
   def date_offset(date = date_string)
