+++ test/helper_test.rb
@@ -1,6 +1,8 @@
 require 'simplecov'
-SimpleCov.start
 require 'minitest'
 require 'minitest/autorun'
 require 'minitest/pride'
-require 'pry'
+
+SimpleCov.start do
+  add_filter "../*.rb"
+end
