+++ test/key_gen_test.rb
@@ -1,6 +1,6 @@
+require './test/helper_test'
 require 'minitest/autorun'
 require 'minitest/pride'
-require 'pry'
 require 'Date'
 require 'Time'
 require './lib/key_gen'
@@ -16,14 +16,14 @@ class KeyGenTest < Minitest::Test
   def test_it_normalizes_key_passed_as_string
     k = KeyGen.new
 
-    assert_instance_of Array, enigma.key_normalizer("12345")
-    assert_equal [12, 23, 34, 45], enigma.key_normalizer("12345")
+    assert_instance_of Array, k.key_normalizer("12345")
+    assert_equal [12, 23, 34, 45], k.key_normalizer("12345")
   end
 
   def test_it_can_create_date_string
-    enigma = Enigma.new
+    k = KeyGen.new
 
-    assert_equal 6, enigma.date_string.length
+    assert_equal 6, k.date_string.length
   end
 
   def test_it_can_create_date_offset
